E2E: Make sure downloads folder existis before trashing it

+++ cypress/support/tasks.ts
@@ -297,9 +297,13 @@ export const insertHslTimetablesDatasetToDb = (
 
 export const emptyDownloadsFolder = () => {
   const downloadsFolder = 'downloads';
-  fs.readdirSync(downloadsFolder).forEach((file) => {
-    deleteFile(`${downloadsFolder}/${file}`);
-  });
+
+  if (fs.existsSync(downloadsFolder)) {
+    fs.readdirSync(downloadsFolder).forEach((file) => {
+      deleteFile(`${downloadsFolder}/${file}`);
+    });
+  }
+
   return true;
 };
